requirements.txt
fastapi==0.120.1
matplotlib==3.10.7
numpy==2.3.4
pandas==2.3.3
psycopg2==2.9.11
python-dotenv==1.2.1
SQLAlchemy==2.0.44
